decorador
