+++ pipelines/datasets/mundo_transfermarkt_competicoes/utils.py
@@ -34,43 +34,18 @@ from bs4 import BeautifulSoup
 import requests
 import numpy as np
 import pandas as pd
+import asyncio
+from pipelines.datasets.mundo_transfermarkt_competicoes.constants import (
+    constants as mundo_constants,
+)
+from pipelines.utils.utils import log
+from pipelines.datasets.mundo_transfermarkt_competicoes.decorators import retry
 
-# import asyncio
-
-"""
-def retry(content_function):
-    async def wrapper(url, attempts=2, wait_time=2, **kwargs):
-        content = None
-        count = 0
-
-        while content is None and count < attempts:
-            # print(f'Attempt {count + 1} of {attempts}')
-            headers = {"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.87 Safari/537.36"}
-            try:
-                response = await asyncio.to_thread(
-                    requests.get, url, headers=headers, timeout=100
-                )
-            except Exception:
-                return None
-            await asyncio.sleep(wait_time)
-            soup = BeautifulSoup(response.text, "html.parser")
-            try:
-                content = content_function(soup, **kwargs)
-            except Exception:
-                if count == (attempts - 1):
-                    # Could not get content
-                    content = None
-            count += 1
-            # await asyncio.sleep(10)
-
-        return content
-
-    return wrapper
 
 @retry
 def get_content(link_soup):
     content = link_soup.find("div", id="main")
-    return content"""
+    return content
 
 
 def process_basico(df, content):
@@ -323,4 +298,226 @@ def vazio(df):
         "afk": None,
     }
     df = df.append(null_content, ignore_index=True)
+    return df
+
+
+async def execucao_coleta():
+    """
+    Execute the program
+    """
+    # Armazena informações do site em um único dataframe.
+    base_url = "https://www.transfermarkt.com/campeonato-brasileiro-serie-a/gesamtspielplan/wettbewerb/BRA1?saison_id={season}&spieltagVon=1&spieltagBis=38"
+    headers = {
+        "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.87 Safari/537.36"
+    }
+    base_link = "https://www.transfermarkt.com"
+    links = []
+
+    ht_tag = []
+    at_tag = []
+    result_tag = []
+
+    ht = []
+    at = []
+    fthg = []
+    ftag = []
+    col_home = []
+    col_away = []
+
+    # expressão regular para encontrar o número de gols marcados pelo time da casa
+    pattern_fthg = re.compile(r"\d:")
+    # expressão regular para encontrar o número de gols marcados pelo time visitante
+    pattern_ftag = re.compile(r":\d")
+
+    # para armazenar os dados das partidas
+    df = pd.DataFrame(
+        {"ht": [], "at": [], "fthg": [], "ftag": [], "col_home": [], "col_away": []}
+    )
+    df_valor = pd.DataFrame({})
+
+    # season = data_atual.year - 1
+    season = mundo_constants.SEASON.value
+    # Pegar o link das partidas
+    # Para cada temporada, adiciona os links dos jogos em `links`
+    log(f"Obtendo links: temporada {season}")
+    site_data = requests.get(base_url.format(season=season), headers=headers)
+    soup = BeautifulSoup(site_data.content, "html.parser")
+    link_tags = soup.find_all("a", attrs={"class": "ergebnis-link"})
+    for tag in link_tags:
+        links.append(re.sub(r"\s", "", tag["href"]))
+    if len(links) % 10 != 0:
+        num_excess = len(links) % 10
+        del links[-num_excess:]
+    tabela = soup.findAll("div", class_="box")
+    for i in range(1, int(len(links) / 10) + 1):
+        for row in tabela[i].findAll("tr"):  # para tudo que estiver em <tr>
+            cells = row.findAll("td")  # variável para encontrar <td>
+            if len(cells) == 7:
+                ht_tag.append(cells[2].findAll(text=True))  # iterando sobre cada linha
+                result_tag.append(
+                    cells[4].findAll(text=True)
+                )  # iterando sobre cada linha
+                at_tag.append(cells[6].findAll(text=True))  # iterando sobre cada linha
+
+    for time in range(len(links)):
+        ht.append(str(ht_tag[time][2]))
+        col_home.append(str(ht_tag[time][0]))
+    for time in range(len(links)):
+        at.append(str(at_tag[time][0]))
+        col_away.append(str(at_tag[time][2]))
+    for tag in result_tag:
+        fthg.append(str(pattern_fthg.findall(str(tag))))
+        ftag.append(str(pattern_ftag.findall(str(tag))))
+
+    # links das estatísticas
+    links_esta = []
+    # links das escalações de cada partida
+    links_valor = []
+
+    for link in links:
+        esta = link.replace("index", "statistik")
+        links_esta.append(esta)
+    for link in links:
+        valor = link.replace("index", "aufstellung")
+        links_valor.append(valor)
+
+    n_links = len(links)
+    log(f"Encontrados {n_links} partidas.")
+    log("Extraindo dados...")
+    for n, link in enumerate(links_esta):
+        # Tentativas de obter os links
+        content = await get_content(base_link + link)
+        if content:
+            try:
+                df = process(df, content)
+            except Exception:
+                try:
+                    df = process_basico(df, content)
+                except Exception:
+                    df = vazio(df)
+        else:
+            df = vazio(df)
+        log(f"{n+1} dados de {n_links} extraídos.")
+    for n, link in enumerate(links_valor):
+        # Tentativas de obter os links
+        content = await get_content(base_link + link)
+        if content:
+            try:
+                df_valor = pegar_valor(df_valor, content)
+            except Exception:
+                try:
+                    df_valor = pegar_valor_sem_tecnico(df_valor, content)
+                except Exception:
+                    df_valor = valor_vazio(df_valor)
+        else:
+            df_valor = valor_vazio(df_valor)
+        log(f"{n+1} valores de {n_links} extraídos.")
+
+    df["ht"] = ht
+    df["at"] = at
+    df["fthg"] = fthg
+    df["ftag"] = ftag
+    df["col_home"] = col_home
+    df["col_away"] = col_away
+
+    # limpar variaveis
+    df["fthg"] = df["fthg"].map(lambda x: x.replace("['", ""))
+    df["fthg"] = df["fthg"].map(lambda x: x.replace(":']", ""))
+
+    df["ftag"] = df["ftag"].map(lambda x: x.replace("[':", ""))
+    df["ftag"] = df["ftag"].map(lambda x: x.replace("']", ""))
+
+    df["col_home"] = df["col_home"].map(lambda x: x.replace("(", ""))
+    df["col_home"] = df["col_home"].map(lambda x: x.replace(".)", ""))
+
+    df["col_away"] = df["col_away"].map(lambda x: x.replace("(", ""))
+    df["col_away"] = df["col_away"].map(lambda x: x.replace(".)", ""))
+
+    df["htag"] = df["htag"].map(lambda x: str(x).replace(")", ""))
+    df["hthg"] = df["hthg"].map(lambda x: str(x).replace("(", ""))
+
+    df_valor["idade_media_titular_vis"] = df_valor["idade_media_titular_vis"].map(
+        lambda x: str(x).replace(" ", "")
+    )
+    df_valor["idade_media_titular_man"] = df_valor["idade_media_titular_man"].map(
+        lambda x: str(x).replace(" ", "")
+    )
+
+    df_valor["valor_equipe_titular_man"] = df_valor["valor_equipe_titular_man"].map(
+        lambda x: str(x).replace("m", "0000")
+    )
+    df_valor["valor_equipe_titular_man"] = df_valor["valor_equipe_titular_man"].map(
+        lambda x: str(x).replace("Th.", "000")
+    )
+    df_valor["valor_equipe_titular_man"] = df_valor["valor_equipe_titular_man"].map(
+        lambda x: str(x).replace(".", "")
+    )
+
+    df_valor["valor_equipe_titular_vis"] = df_valor["valor_equipe_titular_vis"].map(
+        lambda x: str(x).replace("m", "0000")
+    )
+    df_valor["valor_equipe_titular_vis"] = df_valor["valor_equipe_titular_vis"].map(
+        lambda x: str(x).replace("Th.", "000")
+    )
+    df_valor["valor_equipe_titular_vis"] = df_valor["valor_equipe_titular_vis"].map(
+        lambda x: str(x).replace(".", "")
+    )
+
+    df["publico_max"] = df["publico_max"].map(lambda x: str(x).replace(".", ""))
+    df["publico"] = df["publico"].map(lambda x: str(x).replace(".", ""))
+
+    df["test"] = df["publico_max"].replace(to_replace=r"\d", value=1, regex=True)
+
+    def sem_info(x, y):
+        if x == 1:
+            return y
+        return None
+
+    df["test2"] = df.apply(lambda x: sem_info(x["test"], x["publico_max"]), axis=1)
+    df["publico_max"] = df["test2"]
+    del df["test2"]
+    del df["test"]
+
+    df["data"] = pd.to_datetime(df["data"]).dt.date
+    df["horario"] = pd.to_datetime(df["horario"]).dt.strftime("%H:%M")
+
+    df.fillna("", inplace=True)
+
+    df["rodada"] = df["rodada"].astype(np.int64)
+
+    # renomear colunas
+    df = df.rename(
+        columns={
+            "ht": "time_man",
+            "at": "time_vis",
+            "fthg": "gols_man",
+            "ftag": "gols_vis",
+            "col_home": "colocacao_man",
+            "col_away": "colocacao_vis",
+            "ac": "escanteios_vis",
+            "hc": "escanteios_man",
+            "adef": "defesas_vis",
+            "hdef": "defesas_man",
+            "af": "faltas_vis",
+            "afk": "chutes_bola_parada_vis",
+            "aimp": "impedimentos_vis",
+            "as": "chutes_vis",
+            "asofft": "chutes_fora_vis",
+            "hf": "faltas_man",
+            "hfk": "chutes_bola_parada_man",
+            "himp": "impedimentos_man",
+            "hs": "chutes_man",
+            "hsofft": "chutes_fora_man",
+            "htag": "gols_1_tempo_vis",
+            "hthg": "gols_1_tempo_man",
+        }
+    )
+
+    df = pd.concat([df, df_valor], axis=1)
+
+    df["data"] = pd.to_datetime(df["data"])
+    df["ano_campeonato"] = mundo_constants.DATA_ATUAL_ANO.value
+
+    df = df[mundo_constants.ORDEM_COLUNA_FINAL.value]
+
     return df
